Update Arduino library
Instead of hard-coding Serial, now it is possible to specify the Stream object. This allows using this library with any serial port on the Arduino Mega or Teensy, as well as using alternate interfaces like Soft$

A simple Xor checksum has been added to the messages, as well as the function that checks if the checksum is correct.

Instead of returning a pointer to a static local array as a way of returning a whole array, a more safe method is used. getMessage() and readSetPoints() take the address of the array as an argument and the array is defined in the function callers.

Functions have been renamed to comply with the Arduino API style guide (using read and write in their names to make clear what they do).